Okay, first, fixes. (Thoughts inside!)
Apparently voidcrew made a version of the NTSV Bubble for combat, and it was just.. dropped in a direct-to-master commit?
Apparently striketester made this modification (nano_bead) and I just heard nothing about it.  There's also a suspiciously similarly laid out ship in the thunderbird. Counting these with the others, (Shiptest (Pre-And Post Baytiles), Beestation, And Monkestation) How many Bubbles are there? Can we go into the Bubbleverse? An NTSV Bubble No Way Home?
Iunno, but it's still amusing to think my dinky little creation's spread so far, with or without my help.