Adds cargo pods
I have tested NONE of this and MUCH of it is copy pasted, it's likely this does NOT work.

Adds a Cargo computer which stores a list of items a ship is purchasing. Linked to the ship, it uses their bank account to make orders. Doing so will drop pod down a large freight container that holds all your items. This will have all crates stored                                                                     in it, which you can (NOT ADDED YET) move crates in and out of.
To do this, you need to (NOT IMPLEMENTED YET) purchase a beacon that will track whhhhere to send the pod down to.